orte rmaps: Generalize reset_usage() to orte_rmaps_base_node_compute_usage()
Moved and renamed the function from a static function in rmaps_base_binding.c
to an exposed function in rmaps_base_support_fns.c as well as
added the option to include mapped procs in addition to bounds procs in
the usage count calculations. These changes enable the use of this code
not only in the binding code, but also in the mapping code.